backend/requirements.txt
# Framework
Django==5.1.4

# API REST
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0

# Banco de dados
mysqlclient==2.2.4

# Variáveis de ambiente
django-environ==0.11.2

# Hashing de senha com bcrypt
bcrypt==4.2.0

# Extensões úteis para desenvolvimento
django-extensions==3.2.3

# Testes
pytest==8.3.4
pytest-django==4.9.0
factory-boy==3.3.1